+++ packages/atomic/src/components/commerce/facets/atomic-commerce-numeric-facet/atomic-commerce-numeric-facet.tsx
@@ -5,6 +5,9 @@ import {
   ProductListingSummaryState,
   SearchSummaryState,
   Summary,
+  Context,
+  ContextState,
+  buildContext,
 } from '@coveo/headless/commerce';
 import {Component, Element, h, Listen, Prop, State} from '@stencil/core';
 import {FocusTargetController} from '../../../../utils/accessibility-utils';
@@ -18,13 +21,12 @@ import {FacetInfo} from '../../../common/facets/facet-common-store';
 import {FacetContainer} from '../../../common/facets/facet-container/facet-container';
 import {FacetGuard} from '../../../common/facets/facet-guard';
 import {FacetHeader} from '../../../common/facets/facet-header/facet-header';
-import {formatHumanReadable} from '../../../common/facets/numeric-facet/formatter';
 import {NumericFacetValueLink} from '../../../common/facets/numeric-facet/value-link';
 import {NumericFacetValuesContainer} from '../../../common/facets/numeric-facet/values-container';
 import {initializePopover} from '../../../common/facets/popover/popover-type';
 import {
+  defaultCurrencyFormatter,
   defaultNumberFormatter,
-  NumberFormatter,
 } from '../../../common/formats/format-common';
 import {CommerceBindings as Bindings} from '../../atomic-commerce-interface/atomic-commerce-interface';
 import type {Range} from '../facet-number-input/atomic-commerce-facet-number-input';
@@ -45,8 +47,6 @@ export class AtomicCommerceNumericFacet
   @InitializeBindings() public bindings!: Bindings;
   @Element() private host!: HTMLElement;
 
-  @State() private range?: Range;
-
   @BindStateToController('facet')
   @State()
   public facetState!: NumericFacetState;
@@ -55,10 +55,12 @@ export class AtomicCommerceNumericFacet
   @State()
   public summaryState!: SearchSummaryState | ProductListingSummaryState;
 
+  public context!: Context;
+  @BindStateToController('context') contextState!: ContextState;
+
   @State() public error!: Error;
 
   private manualRanges: (NumericRangeRequest & {label?: string})[] = [];
-  private formatter: NumberFormatter = defaultNumberFormatter;
 
   /**
    * The Summary controller instance.
@@ -86,34 +88,41 @@ export class AtomicCommerceNumericFacet
     return this.headerFocus;
   }
 
+  private unsubscribeFacetController!: () => void;
+
   public initialize() {
-    if (!this.facetState) {
+    if (!this.facet) {
       return;
     }
+
+    this.context = buildContext(this.bindings.engine);
+
+    this.unsubscribeFacetController = this.facet.subscribe(
+      () => (this.facetState = this.facet.state)
+    );
+
     this.registerFacetToStore();
   }
 
+  public disconnectedCallback() {
+    this.unsubscribeFacetController();
+  }
+
+  private get formatter() {
+    if (this.field === 'ec_price' || this.field === 'ec_promo_price') {
+      return defaultCurrencyFormatter(this.contextState.currency);
+    }
+    return defaultNumberFormatter;
+  }
+
   private registerFacetToStore() {
-    const {facetId, field} = this.facetState;
+    const {facetId} = this.facetState;
     const facetInfo: FacetInfo = {
       label: () => this.bindings.i18n.t(this.displayName),
       facetId: facetId,
       element: this.host,
       isHidden: () => this.isHidden,
     };
-
-    this.bindings.store.registerFacet('numericFacets', {
-      ...facetInfo,
-      format: (value) =>
-        formatHumanReadable({
-          facetValue: value,
-          logger: this.bindings.engine.logger,
-          i18n: this.bindings.i18n,
-          field: field,
-          manualRanges: this.manualRanges,
-          formatter: this.formatter,
-        }),
-    });
 
     initializePopover(this.host, {
       ...facetInfo,
@@ -122,16 +131,17 @@ export class AtomicCommerceNumericFacet
     });
   }
 
-  @Listen('atomic/numberFormat')
-  public setFormat(event: CustomEvent<NumberFormatter>) {
-    event.preventDefault();
-    event.stopPropagation();
-    this.formatter = event.detail;
-  }
-
   @Listen('atomic/numberInputApply')
   public applyNumberInput({detail}: CustomEvent<Range>) {
-    this.range = {start: detail.start, end: detail.end};
+    const {start, end} = detail;
+    this.facet.setRanges([
+      {
+        start,
+        end,
+        endInclusive: true,
+        state: 'selected',
+      },
+    ]);
   }
 
   public render() {
@@ -154,7 +164,6 @@ export class AtomicCommerceNumericFacet
               onClearFilters={() => {
                 this.focusTarget.focusAfterSearch();
                 this.facet.deselectAll();
-                this.facet.setRanges([]);
               }}
               numberOfActiveValues={this.numberOfSelectedValues}
               isCollapsed={this.isCollapsed}
@@ -169,7 +178,7 @@ export class AtomicCommerceNumericFacet
                   bindings={this.bindings}
                   label={this.displayName}
                   facet={this.facet}
-                  range={this.range}
+                  range={this.facetState.manualRange}
                 ></atomic-commerce-facet-number-input>
               ),
             ]}
@@ -213,10 +222,6 @@ export class AtomicCommerceNumericFacet
   }
 
   private get numberOfSelectedValues() {
-    if (this.range) {
-      return 1;
-    }
-
     return (
       this.facetState.values.filter(({state}) => state === 'selected').length ||
       0
@@ -224,7 +229,7 @@ export class AtomicCommerceNumericFacet
   }
 
   private get hasInputRange() {
-    return !!this.range;
+    return !!this.facetState.manualRange || this.summaryState.isLoading;
   }
 
   private get shouldRenderValues() {