+++ packages/headless/src/features/commerce/facets/facet-set/facet-set-slice.ts
@@ -19,6 +19,7 @@ import {convertToNumericRangeRequests} from '../../../facets/range-facets/numeri
 import {setContext, setView} from '../../context/context-actions';
 import {restoreProductListingParameters} from '../../product-listing-parameters/product-listing-parameters-actions';
 import {fetchProductListing} from '../../product-listing/product-listing-actions';
+import {fetchQuerySuggestions} from '../../query-suggest/query-suggest-actions';
 import {restoreSearchParameters} from '../../search-parameters/search-parameters-actions';
 import {executeSearch} from '../../search/search-actions';
 import '../category-facet/category-facet-actions';
@@ -43,6 +44,7 @@ import {executeCommerceFieldSuggest} from '../facet-search-set/commerce-facet-se
 import {
   toggleExcludeNumericFacetValue,
   toggleSelectNumericFacetValue,
+  updateManualNumericFacetRange,
   updateNumericFacetValues,
 } from '../numeric-facet/numeric-facet-actions';
 import {
@@ -78,10 +80,18 @@ export const commerceFacetSetReducer = createReducer(
     builder
       .addCase(fetchProductListing.fulfilled, handleQueryFulfilled)
       .addCase(executeSearch.fulfilled, handleQueryFulfilled)
-      .addCase(
-        executeCommerceFieldSuggest.fulfilled,
-        handleFieldSuggestionsFulfilled
+      .addCase(executeCommerceFieldSuggest.fulfilled, (state, action) =>
+        handleFieldSuggestionsFulfilled(state, action.payload.facetId)
       )
+      .addCase(fetchQuerySuggestions.fulfilled, (state, action) => {
+        if (!action.payload.fieldSuggestionsFacets) {
+          return;
+        }
+
+        for (const {facetId} of action.payload.fieldSuggestionsFacets) {
+          handleFieldSuggestionsFulfilled(state, facetId);
+        }
+      })
       .addCase(toggleSelectFacetValue, (state, action) => {
         const {facetId, selection} = action.payload;
         const facetRequest = state[facetId]?.request;
@@ -407,6 +417,16 @@ export const commerceFacetSetReducer = createReducer(
 
         handleDeselectAllFacetValues(request);
       })
+      .addCase(updateManualNumericFacetRange, (state, action) => {
+        const {facetId} = action.payload;
+        const request = state[facetId]?.request;
+
+        if (!request) {
+          return;
+        }
+
+        handleDeselectAllFacetValues(request);
+      })
       .addCase(clearAllCoreFacets, setAllFacetValuesToIdle)
       .addCase(setContext, clearAllFacetValues)
       .addCase(setView, clearAllFacetValues)
@@ -458,10 +478,8 @@ function handleQueryFulfilled(
 
 function handleFieldSuggestionsFulfilled(
   state: WritableDraft<CommerceFacetSetState>,
-  action: ReturnType<typeof executeCommerceFieldSuggest.fulfilled>
+  facetId: string
 ) {
-  const facetId = action.payload.facetId;
-
   let facetRequest = state[facetId]?.request;
   if (!facetRequest) {
     state[facetId] = {request: {} as AnyFacetRequest};