+++ packages/headless/src/index.ts
@@ -284,6 +284,12 @@ export type {
 export {buildTab} from './controllers/tab/headless-tab';
 
 export type {
+  TabManagerState,
+  TabManager,
+} from './controllers/tab-manager/headless-tab-manager';
+export {buildTabManager} from './controllers/tab-manager/headless-tab-manager';
+
+export type {
   FacetManagerPayload,
   FacetManagerState,
   FacetManager,
@@ -646,4 +652,8 @@ export type {
 
 export * from './utils/query-expression/query-expression';
 
-export type {GeneratedAnswerFeedback} from './features/generated-answer/generated-answer-analytics-actions';
+export type {
+  GeneratedAnswerFeedback,
+  GeneratedAnswerFeedbackV2,
+  GeneratedAnswerFeedbackOption,
+} from './features/generated-answer/generated-answer-analytics-actions';