+++ packages/headless/src/state/state-sections.ts
@@ -10,6 +10,7 @@ import {CommerceContextState} from '../features/commerce/context/context-state';
 import {DidYouMeanState as CommerceDidYouMeanState} from '../features/commerce/did-you-mean/did-you-mean-state';
 import {CommerceFacetSetState} from '../features/commerce/facets/facet-set/facet-set-state';
 import {FieldSuggestionsOrderState} from '../features/commerce/facets/field-suggestions-order/field-suggestions-order-state';
+import {ManualNumericFacetSetState} from '../features/commerce/facets/numeric-facet/manual-numeric-facet-state';
 import {InstantProductsState} from '../features/commerce/instant-products/instant-products-state';
 import {CommercePaginationState} from '../features/commerce/pagination/pagination-state';
 import {ProductListingState} from '../features/commerce/product-listing/product-listing-state';
@@ -39,6 +40,7 @@ import {GeneratedAnswerState} from '../features/generated-answer/generated-answe
 import {HistoryState} from '../features/history/history-state';
 import {InsightConfigurationState} from '../features/insight-configuration/insight-configuration-state';
 import {InsightInterfaceState} from '../features/insight-interface/insight-interface-state';
+import {UserActionsState} from '../features/insight-user-actions/insight-user-actions-state';
 import {InstantResultsState} from '../features/instant-results/instant-results-state';
 import {PaginationState} from '../features/pagination/pagination-state';
 import {OldProductListingState} from '../features/product-listing/old-product-listing-state';
@@ -501,4 +503,15 @@ export interface GeneratedAnswerSection {
    * The properties related to generative question answering.
    */
   generatedAnswer: GeneratedAnswerState;
+}
+
+export interface InsightUserActionSection {
+  /**
+   * The insight user action state.
+   */
+  insightUserAction: UserActionsState;
+}
+
+export interface ManualRangeSection {
+  manualNumericFacetSet: ManualNumericFacetSetState;
 }