+++ packages/quantic/cypress/e2e/default-2/generatedAnswer/generated-answer-selectors.ts
@@ -16,10 +16,10 @@ export interface GeneratedAnswerSelector extends ComponentSelector {
   generatedAnswerContent: () => CypressSelector;
   generatedAnswerContentContainer: () => CypressSelector;
   feedbackModal: () => CypressSelector;
-  feedbackOption: (index: number) => CypressSelector;
   feedbackSubmitButton: () => CypressSelector;
   feedbackCancelButton: () => CypressSelector;
   feedbackDoneButton: () => CypressSelector;
+  feedbackDocumentUrlInput: () => CypressSelector;
   feedbackDetailsInput: () => CypressSelector;
   rephraseButtons: () => CypressSelector;
   rephraseLabel: () => CypressSelector;
@@ -33,6 +33,8 @@ export interface GeneratedAnswerSelector extends ComponentSelector {
   disclaimer: () => CypressSelector;
   toggleCollapseButton: () => CypressSelector;
   generatingMessage: () => CypressSelector;
+  feedbackOption: (value: string, index: number) => CypressSelector;
+  successMessage: () => CypressSelector;
 }
 
 export const GeneratedAnswerSelectors: GeneratedAnswerSelector = {
@@ -83,18 +85,24 @@ export const GeneratedAnswerSelectors: GeneratedAnswerSelector = {
       '[data-cy="generated-answer__content-container"]'
     ),
   feedbackModal: () => cy.get('lightning-modal'),
-  feedbackOption: (index: number) =>
-    cy.get('lightning-modal').find('lightning-radio-group input').eq(index),
   feedbackSubmitButton: () =>
-    cy.get('lightning-modal').find('[data-cy="feedback-modal-footer__submit"]'),
+    GeneratedAnswerSelectors.feedbackModal().find(
+      '[data-cy="feedback-modal-qna-footer__submit"]'
+    ),
   feedbackCancelButton: () =>
-    cy.get('lightning-modal').find('[data-cy="feedback-modal-footer__cancel"]'),
+    GeneratedAnswerSelectors.feedbackModal().find(
+      '[data-cy="feedback-modal-footer__cancel"]'
+    ),
   feedbackDoneButton: () =>
-    cy.get('lightning-modal').find('[data-cy="feedback-modal-footer__done"]'),
+    GeneratedAnswerSelectors.feedbackModal().find(
+      '[data-cy="feedback-modal-footer__done"]'
+    ),
+  feedbackDocumentUrlInput: () =>
+    GeneratedAnswerSelectors.feedbackModal().find('input[name="documentUrl"]'),
   feedbackDetailsInput: () =>
-    cy
-      .get('lightning-modal')
-      .find('[data-cy="feedback-modal-body__details-input"] textarea'),
+    GeneratedAnswerSelectors.feedbackModal().find(
+      '[data-name="details"] textarea'
+    ),
   rephraseButtons: () =>
     GeneratedAnswerSelectors.get().find(
       '[data-cy="generated-answer__rephrase-buttons"]'
@@ -149,5 +157,13 @@ export const GeneratedAnswerSelectors: GeneratedAnswerSelector = {
   generatingMessage: () =>
     GeneratedAnswerSelectors.get().find(
       'div.generated-answer__collapse-generating-message'
+    ),
+  feedbackOption: (value: string, index: number) =>
+    GeneratedAnswerSelectors.feedbackModal()
+      .find(`.feedback-modal-qna__question input[value=${value}]`)
+      .eq(index),
+  successMessage: () =>
+    GeneratedAnswerSelectors.feedbackModal().find(
+      '[data-cy="feedback-modal-body__success-message"]'
     ),
 };